add MENGINE_CODE_FILE MENGINE_CODE_FUNCTION MENGINE_CODE_LINE rework Platform getDateTime add Logger time stamp add build msvc16_sdl

+++ src/Systems/DX9RenderSystem/DX9RenderProgramVariable.cpp
@@ -33,7 +33,7 @@ namespace Mengine
         void makeVariableData( ProgramVariableDesc & _variable, Vector<T> & _container, uint32_t _type, T * _values, uint32_t _size, uint32_t _count )
         {
             _variable.type = _type;
-            _variable.offset = _container.size();
+            _variable.offset = (uint32_t)_container.size();
             _variable.size = _size;
             _variable.count = _count;
